% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3_bootstrap.R
\name{boot_ttest}
\alias{boot_ttest}
\title{Calculate bootstrap statistics for difference-of-means t-test}
\usage{
boot_ttest(x, y, alpha = 0.05, R = 1000)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}

\item{alpha}{significance level. used to calculate confidence intervals}

\item{R}{number of replications}
}
\value{
\itemize{
   \item{inputs: }{list containing user inputs}
   \item{boot: }{list containing results of bootstrapping, indices used for the pooled sample of x and y, t-statistics for every sample drawn, bias, standard error, jackknife-after-bootstrap}
   \item{CI: }{returned if R >= 200. Lower and upper percentile confidence intervals.}
}
}
\description{
This function calculates R bootstrapped t-test statistics between x and y. It also calculates a variety of statistics such as percentile confidence intervals, bias, standard error and jackknife-after-bootstrap estimates.
}
\seealso{
Rizzo, Maria L. "Statistical Computing with R". Chapman and Hall/CRC, 2007. Chapter 7
}
