% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3_power.R
\name{emp_power}
\alias{emp_power}
\title{Empirical power of a t-test}
\usage{
emp_power(ref_mu, n, mu, sd, alpha = 0.05, R = 1000, ...)
}
\arguments{
\item{ref_mu}{numeric. This is a value indicating the true value of the mean (or difference in means if you are performing a two sample test). See ?t.test() for more information}

\item{n}{numeric. sample size}

\item{mu}{mean value used to draw random normal samples. See ?rnorm() for more information}

\item{sd}{standard deviation used to draw random normal samples}

\item{alpha}{significance level}

\item{R}{number of replications}

\item{...}{optional arguments. You may pass selected parameters from the t.test() function.}
}
\value{
list containing:
\itemize{
  \item{inputs: }{list. user input values}
  \item{data: }{matrix. values of R samples of size n with mean mu and standard deviation sd drawn from a normal distribution using rnorm()}
  \item{p_values: }{vector. resulting p-values of R tests comparing the samples against the reference mean ref_mu}
  \item{power: vector. power calculated by taking the proportion of p-values for which the value falls below or is equal to the significance level alpha}
}
}
\description{
This function takes a reference mean value, sample size, mean, standard deviation, significance level and computes the empirical power of R randomly drawn normally distributed samples with mean mu, standard deviation sd and sample size n. It then calculates R t.tests with reference to the
}
\seealso{
Rizzo, Maria L. 'Statistical Computing with R. Chapman and Hall/CRC, 2007'. (pp. 167-169)
}
