% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a2_linearmodel.R
\name{linear_model}
\alias{linear_model}
\title{Fit a linear regression model}
\usage{
linear_model(formula, data)
}
\arguments{
\item{formula}{an object of class 'formula' or a string that can be coerced to this class. The formula is a symbolic description of the data and should be passed as 'DV ~ IV1 + IV2 + ...' (where DV stands for 'dependent variable' and the IVs are the independent variables). You are also allowed to pass 'DV ~ .' indicating that you will use all all variables in the data as predictors.}

\item{data}{data frame containing the variables of interest}
}
\value{

}
\description{
Use linear_model() to fit a linear regression model. The function computes the QR decomposition and returns this as part of the output list.
}
\seealso{

}
\author{
Jasper Ginn
}
