% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1.R
\name{t_test}
\alias{t_test}
\title{Perform a two-tailed t-test between two variables}
\usage{
t_test(x, y, variance_equal = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}

\item{variance_equal}{boolean (TRUE/FALSE) indicating if variance of x and y are equal or not. If variances are equal, the function will calculate the pooled variance. Else, it will use the formula for unequal variances.}
}
\value{
list of class 't_test' containing input variables, summary statistics & test statistics
}
\description{
Perform a two-tailed t-test between two variables
}
\author{
Jasper Ginn
}
